Update after lab1 and lab2 done

+++ src/main/java/com/capgemini/wsb/fitnesstracker/user/internal/UserServiceImpl.java
@@ -1,12 +1,15 @@
 package com.capgemini.wsb.fitnesstracker.user.internal;
 
 import com.capgemini.wsb.fitnesstracker.user.api.User;
+import com.capgemini.wsb.fitnesstracker.user.api.UserNotFoundException;
 import com.capgemini.wsb.fitnesstracker.user.api.UserProvider;
 import com.capgemini.wsb.fitnesstracker.user.api.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,6 +31,9 @@ class UserServiceImpl implements UserService, UserProvider {
 
     @Override
     public Optional<User> getUser(final Long userId) {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new UserNotFoundException(userId);
+        }
         return userRepository.findById(userId);
     }
 
@@ -42,27 +48,38 @@ class UserServiceImpl implements UserService, UserProvider {
     }
 
     @Override
-    public List<User> getUsersOlderThanSpecifiedAge(final int age) {
-        return userRepository.findAll().stream()
-                .filter(user -> user.getAge() > age)
-                .toList();
+    public List<User> getUsersOlderThan(final LocalDate date) {
+        return userRepository.findAllByDateOfBirthBefore(date);
     }
 
     @Override
     public User updateUser(final Long userId, final User user) {
-        log.info("Updating User with ID {}", userId);
         if (userRepository.findById(userId).isEmpty()) {
-            log.info("No existing User with id {} has been found, creating new User", userId);
-            return createUser(user);
+            throw new UserNotFoundException(userId);
         }
+        log.info("Updating User with ID {}", userId);
+
+        User existingUser = userRepository.findById(userId).get();
+        String firstName = Optional.ofNullable(user.getFirstName()).orElse(existingUser.getFirstName());
+        String lastName = Optional.ofNullable(user.getLastName()).orElse(existingUser.getLastName());
+        LocalDate birthdate = Optional.ofNullable(user.getBirthdate()).orElse(existingUser.getBirthdate());
+        String email = Optional.ofNullable(user.getEmail()).orElse(existingUser.getEmail());
+
         user.setId(userId);
+        user.setFirstName(firstName);
+        user.setLastName(lastName);
+        user.setBirthdate(birthdate);
+        user.setEmail(email);
+
         return userRepository.save(user);
     }
 
     @Override
     public void deleteUser(final Long userId) {
+        if (userRepository.findById(userId).isEmpty()) {
+            throw new UserNotFoundException(userId);
+        }
         log.info("Deleting User with ID {}", userId);
         userRepository.delete(userRepository.findById(userId).get());
     }
-
 }
